Separate managers as local and remote so that we can watch two different clusters and not have to do periodic syncs which could place unnecessary load on the central cluster
