+++ cmd/remote/remote.go
@@ -1,0 +1,79 @@
+/*
+Copyright 2020 The Crossplane Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package remote
+
+import (
+	"time"
+
+	"github.com/crossplane/crossplane/apis/apiextensions"
+	"github.com/pkg/errors"
+	crds "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
+	"k8s.io/client-go/rest"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+
+	"github.com/crossplane/crossplane-runtime/pkg/logging"
+
+	"github.com/crossplane/agent/pkg/controllers/cluster/crd"
+	"github.com/crossplane/agent/pkg/controllers/cluster/syncer"
+)
+
+type Setup func(mgr manager.Manager, localClient client.Client, logger logging.Logger) error
+
+type Agent struct {
+	period    time.Duration
+	namespace string
+
+	ClusterConfig *rest.Config
+	DefaultConfig *rest.Config
+}
+
+func (a *Agent) Run(log logging.Logger, period time.Duration) error {
+	log.Debug("Starting", "sync-period", period.String())
+
+	localClient, err := client.New(ctrl.GetConfigOrDie(), client.Options{})
+	if err != nil {
+		return errors.Wrap(err, "cannot create local client")
+	}
+
+	mgr, err := ctrl.NewManager(a.ClusterConfig, ctrl.Options{SyncPeriod: &period})
+	if err != nil {
+		return errors.Wrap(err, "cannot start remote cluster manager")
+	}
+
+	if err := crds.AddToScheme(mgr.GetScheme()); err != nil {
+		return errors.Wrap(err, "Cannot add CustomResourceDefinition API to scheme")
+	}
+
+	if err := apiextensions.AddToScheme(mgr.GetScheme()); err != nil {
+		return errors.Wrap(err, "Cannot add Crossplane apiextensions API to scheme")
+	}
+
+	for _, setup := range []Setup{
+		crd.Setup,
+		syncer.SetupInfraPubSync,
+		syncer.SetupInfraDefSync,
+		syncer.SetupCompositionSync,
+	} {
+		if err := setup(mgr, localClient, log); err != nil {
+			return err
+		}
+	}
+
+	return errors.Wrap(mgr.Start(ctrl.SetupSignalHandler()), "cannot start controller manager")
+}
