+++ pkg/controllers/cluster/syncer/reconciler.go
@@ -24,132 +24,23 @@ import (
 	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	kcontroller "sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/crossplane/crossplane-runtime/pkg/event"
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 	rresource "github.com/crossplane/crossplane-runtime/pkg/resource"
-	"github.com/crossplane/crossplane/apis/apiextensions/v1alpha1"
 	"github.com/crossplane/crossplane/apis/apiextensions/v1alpha1/ccrd"
 
 	"github.com/crossplane/agent/pkg/resource"
 )
 
 const (
-	maxConcurrency = 5
-	timeout        = 2 * time.Minute
-	shortWait      = 30 * time.Second
-	longWait       = 1 * time.Minute
+	timeout   = 2 * time.Minute
+	shortWait = 30 * time.Second
+	longWait  = 1 * time.Minute
 )
-
-// SetupInfraDefSync adds a controller that syncs InfrastructureDefinitions.
-func SetupInfraDefSync(mgr ctrl.Manager, localClient client.Client, log logging.Logger) error {
-	name := "InfrastructureDefinitions"
-
-	nl := func() runtime.Object { return &v1alpha1.InfrastructureDefinitionList{} }
-	gi := func(l runtime.Object) []rresource.Object {
-		list, _ := l.(*v1alpha1.InfrastructureDefinitionList)
-		result := make([]rresource.Object, len(list.Items))
-		for i, val := range list.Items {
-			obj, _ := val.DeepCopyObject().(rresource.Object)
-			result[i] = obj
-		}
-		return result
-	}
-	ni := func() rresource.Object { return &v1alpha1.InfrastructureDefinition{} }
-
-	r := NewReconciler(mgr,
-		WithLogger(log.WithValues("controller", name)),
-		WithRecorder(event.NewAPIRecorder(mgr.GetEventRecorderFor(name))),
-		WithCRDName("infrastructuredefinitions.apiextensions.crossplane.io"),
-		WithNewInstanceFn(ni),
-		WithNewListFn(nl),
-		WithGetItemsFn(gi),
-		WithLocalClient(rresource.ClientApplicator{
-			Client:     localClient,
-			Applicator: rresource.NewAPIUpdatingApplicator(localClient),
-		}))
-
-	return ctrl.NewControllerManagedBy(mgr).
-		Named(name).
-		For(&v1alpha1.InfrastructureDefinition{}).
-		WithOptions(kcontroller.Options{MaxConcurrentReconciles: maxConcurrency}).
-		Complete(r)
-}
-
-// SetupInfraPubSync adds a controller that reconciles ApplicationConfigurations.
-func SetupInfraPubSync(mgr ctrl.Manager, localClient client.Client, log logging.Logger) error {
-	name := "InfrastructurePublications"
-
-	nl := func() runtime.Object { return &v1alpha1.InfrastructurePublicationList{} }
-	gi := func(l runtime.Object) []rresource.Object {
-		list, _ := l.(*v1alpha1.InfrastructurePublicationList)
-		result := make([]rresource.Object, len(list.Items))
-		for i, val := range list.Items {
-			obj, _ := val.DeepCopyObject().(rresource.Object)
-			result[i] = obj
-		}
-		return result
-	}
-	ni := func() rresource.Object { return &v1alpha1.InfrastructurePublication{} }
-
-	r := NewReconciler(mgr,
-		WithLogger(log.WithValues("controller", name)),
-		WithRecorder(event.NewAPIRecorder(mgr.GetEventRecorderFor(name))),
-		WithCRDName("infrastructurepublications.apiextensions.crossplane.io"),
-		WithNewInstanceFn(ni),
-		WithNewListFn(nl),
-		WithGetItemsFn(gi),
-		WithLocalClient(rresource.ClientApplicator{
-			Client:     localClient,
-			Applicator: rresource.NewAPIUpdatingApplicator(localClient),
-		}))
-
-	return ctrl.NewControllerManagedBy(mgr).
-		Named(name).
-		For(&v1alpha1.InfrastructurePublication{}).
-		WithOptions(kcontroller.Options{MaxConcurrentReconciles: maxConcurrency}).
-		Complete(r)
-}
-
-// SetupCompositionSync adds a controller that syncs Compositions.
-func SetupCompositionSync(mgr ctrl.Manager, localClient client.Client, log logging.Logger) error {
-	name := "Compositions"
-
-	nl := func() runtime.Object { return &v1alpha1.CompositionList{} }
-	gi := func(l runtime.Object) []rresource.Object {
-		list, _ := l.(*v1alpha1.CompositionList)
-		result := make([]rresource.Object, len(list.Items))
-		for i, val := range list.Items {
-			obj, _ := val.DeepCopyObject().(rresource.Object)
-			result[i] = obj
-		}
-		return result
-	}
-	ni := func() rresource.Object { return &v1alpha1.Composition{} }
-
-	r := NewReconciler(mgr,
-		WithLogger(log.WithValues("controller", name)),
-		WithRecorder(event.NewAPIRecorder(mgr.GetEventRecorderFor(name))),
-		WithCRDName("compositions.apiextensions.crossplane.io"),
-		WithNewInstanceFn(ni),
-		WithNewListFn(nl),
-		WithGetItemsFn(gi),
-		WithLocalClient(rresource.ClientApplicator{
-			Client:     localClient,
-			Applicator: rresource.NewAPIUpdatingApplicator(localClient),
-		}))
-
-	return ctrl.NewControllerManagedBy(mgr).
-		Named(name).
-		For(&v1alpha1.Composition{}).
-		WithOptions(kcontroller.Options{MaxConcurrentReconciles: maxConcurrency}).
-		Complete(r)
-}
 
 // ReconcilerOption is used to configure the Reconciler.
 type ReconcilerOption func(*Reconciler)
